requirements.txt
glfw~=2.10.0
pyopengl~=3.1.10